+++ internal/api/response.go
@@ -6,28 +6,70 @@ import (
 	"time"
 )
 
-// NewDLCDataResponse transforms a entity.DLCData to dlcData response
-func NewDLCDataResponse(
+// NewOracleAnnouncement converts a DLCData structure to an oracle announcement
+func NewOracleAnnouncement(
 	oraclePubKey *dlccrypto.SchnorrPublicKey,
-	dlcData *entity.DLCData) *DLCDataResponse {
-	return &DLCDataResponse{
-		OraclePublicKey: oraclePubKey.EncodeToString(),
-		PublishedDate:   dlcData.PublishedDate,
-		AssetID:         dlcData.AssetID,
-		Rvalue:          dlcData.Rvalue,
-		Signature:       dlcData.Signature,
-		Value:           dlcData.Value,
+	eventData *entity.EventData) *OracleAnnouncement {
+	descriptor := DecompositionDescriptor{
+		Base:      eventData.Base,
+		IsSigned:  eventData.IsSigned,
+		Unit:      eventData.Unit,
+		Precision: eventData.Precision,
+	}
+	event := OracleEvent{
+		Nonces:          eventData.Nonces,
+		EventMaturity:   eventData.PublishedDate,
+		EventDescriptor: descriptor,
+		EventID:         eventData.GetEventID(),
+	}
+	announcement := OracleAnnouncement{
+		AnnouncementSignature: eventData.AnnouncementSignature,
+		OraclePublicKey:       oraclePubKey.EncodeToString(),
+		OracleEvent:           event,
+	}
+
+	return &announcement
+}
+
+// NewOracleAttestation creates a new OracleAttestation structure from the given eventData
+func NewOracleAttestation(eventData *entity.EventData) *OracleAttestation {
+	return &OracleAttestation{
+		EventID:    eventData.GetEventID(),
+		Signatures: eventData.Signatures,
+		Values:     eventData.Values,
 	}
 }
 
-// DLCDataResponse represents the DLC data struct sent by AssetController
-type DLCDataResponse struct {
-	OraclePublicKey string    `json:"oraclePublicKey"`
-	PublishedDate   time.Time `json:"publishDate"`
-	AssetID         string    `json:"asset"`
-	Rvalue          string    `json:"rvalue"`
-	Signature       string    `json:"signature,omitempty"`
-	Value           string    `json:"value,omitempty"`
+// DecompositionDescriptor contains information to about an event using
+// numerical decomposition
+type DecompositionDescriptor struct {
+	Base      int    `json:"base"`
+	IsSigned  bool   `json:"isSigned"`
+	Unit      string `json:"unit"`
+	Precision int    `json:"precision"`
+}
+
+// OracleEvent contains information about an event
+type OracleEvent struct {
+	Nonces          []string                `json:"nonces"`
+	EventMaturity   time.Time               `json:"eventMaturity"`
+	EventDescriptor DecompositionDescriptor `json:"eventDescriptor"`
+	EventID         string                  `json:"eventId"`
+}
+
+// OracleAnnouncement contains information about an event and a signature over
+// the OracleEvent structure
+type OracleAnnouncement struct {
+	AnnouncementSignature string      `json:"announcementSignature"`
+	OraclePublicKey       string      `json:"oraclePublicKey"`
+	OracleEvent           OracleEvent `json:"oracleEvent"`
+}
+
+// OracleAttestation contains information about the outcome of an event
+type OracleAttestation struct {
+	EventID    string   `json:"eventId"`
+	Signatures []string `json:"signatures"`
+	Values     []string `json:"values"`
 }
 
 // AssetConfigResponse represents the configuration of an asset api