+++ internal/database/entity/timestamp.go
@@ -2,8 +2,8 @@ package entity
 
 import "time"
 
-// Base entity to add db timestamp field
-type Base struct {
+// Timestamp entity to add db timestamp field
+type Timestamp struct {
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt *time.Time `sql:"index"`