+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Enable decomposition of numerical event outcomes into digits signed separately using different nonces.
+
+### Fixed
+- Issue with concurrent requests for an event that is not yet in the DB.
+
 ## [0.0.4] - 2020-26-10
 
 ### Changed